requirements.txt
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.4
constantly==15.1.0
entrypoints==0.3
flake8==3.7.8
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.0.4
Pillow==6.1.0
PyAutoGUI==0.9.47
pycodestyle==2.5.0
pyflakes==2.1.1
PyGetWindow==0.0.7
PyHamcrest==1.9.0
PyMsgBox==1.0.7
pynput==1.4.2
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PyTweening==1.0.3
PyYAML==5.1.2
selenium==3.141.0
six==1.12.0
Twisted==19.7.0
urllib3==1.25.3
zope.interface==4.6.0
